+++ NEWS.md
@@ -1,3 +1,5 @@
+# evavelo (development version)
+
 # evavelo 0.5.0
 
 - User Interface improvement #33